Clean config (#170)
* Rename config to runner_config to be sure what we are talking about.

* Use driver config memory instead of having a separate runner config MemoryPerJob.

* Move configuration initialization to configloader package.

This cleans up the configuration handling, will allow extended configuration files that allow options other than the runner to be specified, and removes the viper dependency from the agent package.

* Move api_url out of runner config. It is not relevant to the runner, but relevant to all existing taskers.

* Set log15 level based on config.

* Some comment cleanup in configloader.

* Demonstration of loading custom configuration like stats.

* Remove debugging line.

* Fix test build.

+++ drivers/docker/docker.go
@@ -147,7 +147,7 @@ func (drv *DockerDriver) createContainer(task drivers.ContainerTask) (string, er
 		Config: &docker.Config{
 			Env:       envvars,
 			Cmd:       cmd,
-			Memory:    drv.conf.Memory,
+			Memory:    int64(drv.conf.Memory),
 			CPUShares: drv.conf.CPUShares,
 			Hostname:  drv.hostname,
 			Image:     task.Image(),
